UI: Break out the hierarchy functions into ui/root.cpp/h
