+++ CMakeLists.txt
@@ -1040,6 +1040,8 @@ add_library(native STATIC
 	ext/native/thread/threadutil.h
 	ext/native/thread/threadpool.cpp
 	ext/native/thread/threadpool.h
+	ext/native/ui/root.cpp
+	ext/native/ui/root.h
 	ext/native/ui/screen.cpp
 	ext/native/ui/screen.h
 	ext/native/ui/ui.cpp
