+++ ext/native/ui/root.cpp
@@ -1,0 +1,202 @@
+#include <mutex>
+
+#include "base/timeutil.h"
+#include "ui/root.h"
+#include "ui/viewgroup.h"
+
+namespace UI {
+
+static std::mutex focusLock;
+static std::vector<int> focusMoves;
+extern bool focusForced;
+
+void LayoutViewHierarchy(const UIContext &dc, ViewGroup *root) {
+	if (!root) {
+		ELOG("Tried to layout a view hierarchy from a zero pointer root");
+		return;
+	}
+	const Bounds &rootBounds = dc.GetBounds();
+
+	MeasureSpec horiz(EXACTLY, rootBounds.w);
+	MeasureSpec vert(EXACTLY, rootBounds.h);
+
+	// Two phases - measure contents, layout.
+	root->Measure(dc, horiz, vert);
+	// Root has a specified size. Set it, then let root layout all its children.
+	root->SetBounds(rootBounds);
+	root->Layout();
+}
+
+void MoveFocus(ViewGroup *root, FocusDirection direction) {
+	if (!GetFocusedView()) {
+		// Nothing was focused when we got in here. Focus the first non-group in the hierarchy.
+		root->SetFocus();
+		return;
+	}
+
+	NeighborResult neigh(0, 0);
+	neigh = root->FindNeighbor(GetFocusedView(), direction, neigh);
+
+	if (neigh.view) {
+		neigh.view->SetFocus();
+		root->SubviewFocused(neigh.view);
+	}
+}
+
+// TODO: Figure out where this should really live.
+// Simple simulation of key repeat on platforms and for gamepads where we don't
+// automatically get it.
+
+static int frameCount;
+
+// Ignore deviceId when checking for matches. Turns out that Ouya for example sends
+// completely broken input where the original keypresses have deviceId = 10 and the repeats
+// have deviceId = 0.
+struct HeldKey {
+	int key;
+	int deviceId;
+	double triggerTime;
+
+	// Ignores startTime
+	bool operator <(const HeldKey &other) const {
+		if (key < other.key) return true;
+		return false;
+	}
+	bool operator ==(const HeldKey &other) const { return key == other.key; }
+};
+
+static std::set<HeldKey> heldKeys;
+
+const double repeatDelay = 15 * (1.0 / 60.0f);  // 15 frames like before.
+const double repeatInterval = 5 * (1.0 / 60.0f);  // 5 frames like before.
+
+bool KeyEvent(const KeyInput &key, ViewGroup *root) {
+	bool retval = false;
+	// Ignore repeats for focus moves.
+	if ((key.flags & (KEY_DOWN | KEY_IS_REPEAT)) == KEY_DOWN) {
+		if (IsDPadKey(key)) {
+			// Let's only repeat DPAD initially.
+			HeldKey hk;
+			hk.key = key.keyCode;
+			hk.deviceId = key.deviceId;
+			hk.triggerTime = time_now_d() + repeatDelay;
+
+			// Check if the key is already held. If it is, ignore it. This is to avoid
+			// multiple key repeat mechanisms colliding.
+			if (heldKeys.find(hk) != heldKeys.end()) {
+				return false;
+			}
+
+			heldKeys.insert(hk);
+			std::lock_guard<std::mutex> lock(focusLock);
+			focusMoves.push_back(key.keyCode);
+			retval = true;
+		}
+	}
+	if (key.flags & KEY_UP) {
+		// We ignore the device ID here (in the comparator for HeldKey), due to the Ouya quirk mentioned above.
+		if (!heldKeys.empty()) {
+			HeldKey hk;
+			hk.key = key.keyCode;
+			hk.deviceId = key.deviceId;
+			hk.triggerTime = 0.0; // irrelevant
+			if (heldKeys.find(hk) != heldKeys.end()) {
+				heldKeys.erase(hk);
+				retval = true;
+			}
+		}
+	}
+
+	retval = root->Key(key);
+
+	// Ignore volume keys and stuff here. Not elegant but need to propagate bools through the view hierarchy as well...
+	switch (key.keyCode) {
+	case NKCODE_VOLUME_DOWN:
+	case NKCODE_VOLUME_UP:
+	case NKCODE_VOLUME_MUTE:
+		retval = false;
+		break;
+	}
+
+	return retval;
+}
+
+static void ProcessHeldKeys(ViewGroup *root) {
+	double now = time_now_d();
+
+restart:
+
+	for (std::set<HeldKey>::iterator iter = heldKeys.begin(); iter != heldKeys.end(); ++iter) {
+		if (iter->triggerTime < now) {
+			KeyInput key;
+			key.keyCode = iter->key;
+			key.deviceId = iter->deviceId;
+			key.flags = KEY_DOWN;
+			KeyEvent(key, root);
+
+			std::lock_guard<std::mutex> lock(focusLock);
+			focusMoves.push_back(key.keyCode);
+
+			// Cannot modify the current item when looping over a set, so let's do this instead.
+			HeldKey hk = *iter;
+			heldKeys.erase(hk);
+			hk.triggerTime = now + repeatInterval;
+			heldKeys.insert(hk);
+			goto restart;
+		}
+	}
+}
+
+bool TouchEvent(const TouchInput &touch, ViewGroup *root) {
+	focusForced = false;
+	root->Touch(touch);
+	if ((touch.flags & TOUCH_DOWN) && !focusForced) {
+		EnableFocusMovement(false);
+	}
+	return true;
+}
+
+bool AxisEvent(const AxisInput &axis, ViewGroup *root) {
+	root->Axis(axis);
+	return true;
+}
+
+void UpdateViewHierarchy(ViewGroup *root) {
+	ProcessHeldKeys(root);
+	frameCount++;
+
+	if (!root) {
+		ELOG("Tried to update a view hierarchy from a zero pointer root");
+		return;
+	}
+
+	if (focusMoves.size()) {
+		std::lock_guard<std::mutex> lock(focusLock);
+		EnableFocusMovement(true);
+		if (!GetFocusedView()) {
+			View *defaultView = root->GetDefaultFocusView();
+			// Can't focus what you can't see.
+			if (defaultView && defaultView->GetVisibility() == V_VISIBLE) {
+				root->GetDefaultFocusView()->SetFocus();
+			} else {
+				root->SetFocus();
+			}
+			root->SubviewFocused(GetFocusedView());
+		} else {
+			for (size_t i = 0; i < focusMoves.size(); i++) {
+				switch (focusMoves[i]) {
+				case NKCODE_DPAD_LEFT: MoveFocus(root, FOCUS_LEFT); break;
+				case NKCODE_DPAD_RIGHT: MoveFocus(root, FOCUS_RIGHT); break;
+				case NKCODE_DPAD_UP: MoveFocus(root, FOCUS_UP); break;
+				case NKCODE_DPAD_DOWN: MoveFocus(root, FOCUS_DOWN); break;
+				}
+			}
+		}
+		focusMoves.clear();
+	}
+
+	root->Update();
+	DispatchEvents();
+}
+
+}
