+++ ext/native/ui/ui_screen.cpp
@@ -7,6 +7,7 @@
 #include "ui/ui_screen.h"
 #include "ui/ui_context.h"
 #include "ui/screen.h"
+#include "ui/root.h"
 #include "i18n/i18n.h"
 #include "gfx_es2/draw_buffer.h"
