+++ ext/native/ui/viewgroup.h
@@ -397,11 +397,4 @@ private:
 	std::set<int> hidden_;
 };
 
-void LayoutViewHierarchy(const UIContext &dc, ViewGroup *root);
-void UpdateViewHierarchy(ViewGroup *root);
-// Hooks arrow keys for navigation
-bool KeyEvent(const KeyInput &key, ViewGroup *root);
-bool TouchEvent(const TouchInput &touch, ViewGroup *root);
-bool AxisEvent(const AxisInput &axis, ViewGroup *root);
-
 }  // namespace UI
